refactor: styled components 위치 이동
- styled copmonents 위치를 style 디렉터리에서 각 컴포넌트 파일 내부로 이동
