+++ src/components/Head/index.jsx
@@ -1,4 +1,22 @@
-import { HeadContainer, BackButton, Title } from '../../style/head';
+import styled from 'styled-components';
+
+const HeadContainer = styled.div`
+  display: flex;
+  padding: 22px 24px 0;
+  gap: 18px;
+`;
+
+const BackButton = styled.button`
+  background-color: #fff;
+`;
+
+const Title = styled.span`
+  font-style: normal;
+  font-weight: 500;
+  font-size: 16px;
+  line-height: 19px;
+  color: #383838;
+`;
 
 function Head({ title }) {
   return (
