+++ src/components/SubmitButton/index.jsx
@@ -1,4 +1,16 @@
-import { Button } from '../../style/addForm';
+import styled from 'styled-components';
+
+const Button = styled.button`
+  display: ${props => (props.hidden ? 'none' : 'block')};
+  width: ${props => props.width};
+  height: ${props => props.height};
+  font-weight: 700;
+  font-size: 14px;
+  line-height: 16px;
+  text-align: right;
+  color: #04c09e;
+  background-color: white;
+`;
 
 function SubmitButton({ label, width, height, hidden }) {
   return (
