--- src/style/head.js
@@ -1,19 +1,0 @@
-import styled from 'styled-components';
-
-export const HeadContainer = styled.div`
-  display: flex;
-  padding: 22px 24px 0;
-  gap: 18px;
-`;
-
-export const BackButton = styled.button`
-  background-color: #fff;
-`;
-
-export const Title = styled.span`
-  font-style: normal;
-  font-weight: 500;
-  font-size: 16px;
-  line-height: 19px;
-  color: #383838;
-`;
